refactor: extract shared logger & transport
